banking_system/requirements_linux.txt
asgiref==3.8.1
coverage==7.4.4
Django==3.2.18
django-widget-tweaks==1.4.12
joblib==1.2.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.4.0
tzdata==2024.1
